+++ pkg/ddc/alluxio/transform_init_users_test.go
@@ -1,5 +1,4 @@
 /*
-
 Copyright 2023 The Fluid Author.
 
 Licensed under the Apache License, Version 2.0 (the "License");